seperate provider settings
