+++ templates/components/providers/typescript/gemini/provider.ts
@@ -1,0 +1,16 @@
+import {
+  Gemini,
+  GEMINI_EMBEDDING_MODEL,
+  GEMINI_MODEL,
+  GeminiEmbedding,
+} from "@llamaindex/google";
+import { Settings } from "llamaindex";
+
+export function setupProvider() {
+  Settings.llm = new Gemini({
+    model: process.env.MODEL as GEMINI_MODEL,
+  });
+  Settings.embedModel = new GeminiEmbedding({
+    model: process.env.EMBEDDING_MODEL as GEMINI_EMBEDDING_MODEL,
+  });
+}
