+++ templates/components/providers/typescript/mistral/provider.ts
@@ -1,0 +1,16 @@
+import {
+  ALL_AVAILABLE_MISTRAL_MODELS,
+  MistralAI,
+  MistralAIEmbedding,
+  MistralAIEmbeddingModelType,
+} from "@llamaindex/mistral";
+import { Settings } from "llamaindex";
+
+export function setupProvider() {
+  Settings.llm = new MistralAI({
+    model: process.env.MODEL as keyof typeof ALL_AVAILABLE_MISTRAL_MODELS,
+  });
+  Settings.embedModel = new MistralAIEmbedding({
+    model: process.env.EMBEDDING_MODEL as MistralAIEmbeddingModelType,
+  });
+}
