+++ templates/types/streaming/express/package.json
@@ -24,14 +24,6 @@
     "duck-duck-scrape": "^2.2.5",
     "express": "^4.18.2",
     "llamaindex": "^0.9.0",
-    "@llamaindex/readers": "^2.0.0",
-    "@llamaindex/anthropic": "^0.1.0",
-    "@llamaindex/google": "^0.0.7",
-    "@llamaindex/groq": "^0.0.51",
-    "@llamaindex/huggingface": "^0.0.36",
-    "@llamaindex/mistral": "^0.0.5",
-    "@llamaindex/ollama": "^0.0.40",
-    "@llamaindex/openai": "^0.1.52",
     "pdf2json": "^3.0.5",
     "ajv": "^8.12.0",
     "@e2b/code-interpreter": "^1.0.4",
